Introduce TrajectoryDataWriter interface and refactor file writer
